refactor(vc-progress): update
