+++ components/vc-progress/index.ts
@@ -1,5 +1,5 @@
 // based on rc-progress 2.5.2
-import Progress, { Line, Circle } from './src/';
+import Progress, { Line, Circle } from './src';
 
 export { Line, Circle };
