+++ components/vc-progress/src/index.ts
@@ -1,7 +1,8 @@
 import Line from './Line';
 import Circle from './Circle';
+import type { ProgressProps } from './types';
 
-export { Line, Circle };
+export { Line, Circle, ProgressProps };
 
 export default {
   Line,
